+++ publisher/src/components/Settings/AgencySettingsJurisdictions.tsx
@@ -18,6 +18,7 @@
 /* eslint-disable camelcase */
 import addIcon from "@justice-counts/common/assets/add-icon.svg";
 import blackCheck from "@justice-counts/common/assets/black-check-icon.svg";
+import { Button } from "@justice-counts/common/components/Button";
 import mappedJurisdictionsJSONData from "@justice-counts/common/fips_with_county_subdivisions.json";
 import { Jurisdiction } from "@justice-counts/common/types";
 import { observer } from "mobx-react-lite";
@@ -31,7 +32,6 @@ import {
   Checkbox,
   CheckboxWrapper,
 } from "../MetricConfiguration";
-import { Button } from "../shared/Button";
 import { SettingProps } from "./AgencySettings";
 import {
   AddIcon,
